# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/allergie": {
            "get": {
                "tags": [
                    "Allergie pour une commit "
                ],
                "summary": "Get list of allergy",
                "description": "Returns list of allergy",
                "operationId": "getallergieList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No allergy found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Allergie"
                ],
                "summary": "Create a new allergy",
                "description": "Creates a new allergy with the provided data",
                "operationId": "storeAllergie",
                "requestBody": {
                    "description": "Allergie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Allergie label"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Allergie created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/allergie/{id}": {
            "get": {
                "tags": [
                    "Allergie"
                ],
                "summary": "Get a specific allergy by ID",
                "description": "Retrieves an allergy based on its ID",
                "operationId": "getAllergieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the allergy to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Allergie not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Allergie"
                ],
                "summary": "Update an existing allergy",
                "description": "Updates an existing allergy with the provided data",
                "operationId": "updateAllergie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the allergy to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Allergie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Updated Allergie label"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Allergie updated successfully"
                    },
                    "404": {
                        "description": "Allergie not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Allergie"
                ],
                "summary": "Delete an existing allergy",
                "description": "Deletes an existing allergy by its ID",
                "operationId": "deleteAllergie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the allergy to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Allergie deleted successfully"
                    },
                    "404": {
                        "description": "Allergie not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}